requirements.txt
absl-py==2.2.2
accelerate==1.7.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
audioread==3.0.1
blis==1.3.0
blobfile==3.0.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
colour==0.1.5
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
easydict==1.13
einops==0.8.1
executing==2.2.0
ffmpeg==1.4
ffmpy==0.6.0
filelock==3.13.1
fire==0.1.3
fonttools==4.58.0
fsspec==2024.6.1
grpcio==1.71.0
h5py==3.13.0
hf-xet==1.1.2
huggingface-hub==0.32.1
icecream==2.1.4
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
librosa==0.9.2
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpi4py==3.1.4
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.3
numba==0.61.2
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.0.0
pip==25.1.1
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.9
protobuf==6.31.0
psutil==7.0.0
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-minimize==0.0.2
PyYAML==6.0.2
requests==2.32.3
resampy==0.4.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==60.9.3
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smplx==0.1.28
SoundFile==0.10.3.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchcrepe==0.0.24
torchvision==0.18.1+cu121
tqdm==4.45.0
triton==2.3.1
typer==0.15.4
typing_extensions==4.12.2
typing-inspection==0.4.1
Unidecode==1.1.1
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
wrapt==1.17.2
